+++ credit/build.gradle
@@ -1,8 +1,8 @@
 plugins {
-	id 'org.springframework.boot' version '2.4.0'
-	id 'io.spring.dependency-management' version '1.0.10.RELEASE'
-	id 'java'
-	id "io.freefair.lombok" version "5.3.0"
+    id 'org.springframework.boot' version '2.4.0'
+    id 'io.spring.dependency-management' version '1.0.10.RELEASE'
+    id 'java'
+    id "io.freefair.lombok" version "5.3.0"
 
 }
 
@@ -10,27 +10,28 @@ group = 'de.upb.crypto'
 sourceCompatibility = '15'
 
 repositories {
-	mavenCentral()
-	jcenter()
+    mavenCentral()
+    jcenter()
 }
 
 dependencies {
-	implementation 'org.springframework.boot:spring-boot-starter-web'
-	implementation "io.springfox:springfox-boot-starter:3.0.0"
-	implementation 'org.springframework.boot:spring-boot-starter-webflux'
-	testImplementation 'org.springframework.boot:spring-boot-starter-test'
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+    implementation "io.springfox:springfox-boot-starter:3.0.0"
+    implementation 'org.springframework.boot:spring-boot-starter-webflux'
+    testImplementation 'org.springframework.boot:spring-boot-starter-test'
 
-	implementation project(':protocoldefinition')
+    implementation project(':protocoldefinition')
+    implementation project(':cryptoprotocol')
 }
 
 test {
-	useJUnitPlatform()
+    useJUnitPlatform()
 }
 
 bootBuildImage {
-	if (System.getenv('VERSION') == null || System.getenv('VERSION') == '') {
-		imageName = "upbcuk/incentive-service-credit"
-	} else {
-		imageName = "upbcuk/incentive-service-credit:" + System.getenv('VERSION')
-	}
+    if (System.getenv('VERSION') == null || System.getenv('VERSION') == '') {
+        imageName = "upbcuk/incentive-service-credit"
+    } else {
+        imageName = "upbcuk/incentive-service-credit:" + System.getenv('VERSION')
+    }
 }
