+++ credit/src/main/java/de/upb/crypto/incentive/services/credit/CreditApplication.java
@@ -1,4 +1,4 @@
-package de.upv.crypto.incentive.services.credit;
+package de.upb.crypto.incentive.services.credit;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -6,7 +6,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class CreditApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(CreditApplication.class, args);
-	}
+    public static void main(String[] args) {
+        SpringApplication.run(CreditApplication.class, args);
+    }
 }
