+++ credit/src/main/resources/META-INF/additional-spring-configuration-metadata.json
@@ -4,5 +4,6 @@
       "name": "t2.server.gateway",
       "type": "java.lang.String",
       "description": "Basepath at which issuing service api can be reached."
-  }
-] }
+    }
+  ]
+}
